wasi-io: inline view into lib.rs. improve docs.
