+++ crates/wasi-io/src/lib.rs
@@ -19,12 +19,84 @@ pub mod bindings;
 mod impls;
 pub mod poll;
 pub mod streams;
-mod view;
-
-pub use view::{IoImpl, IoView};
 
 #[doc(no_inline)]
 pub use async_trait::async_trait;
+
+use wasmtime::component::ResourceTable;
+
+/// A trait which provides access to the [`ResourceTable`] inside the
+/// embedder's `T` of [`Store<T>`][`Store`].
+///
+/// This crate's WASI Host implementations depend on the contents of
+/// [`ResourceTable`]. The `T` type [`Store<T>`][`Store`] is defined in each
+/// embedding of Wasmtime. These implementations are connected to the
+/// [`Linker<T>`][`Linker`] by the
+/// [`add_to_linker_sync`](crate::add_to_linker_sync) and
+/// [`add_to_linker_async`](crate::add_to_linker_async) functions.
+///
+/// # Example
+///
+/// ```
+/// use wasmtime::{Config, Engine};
+/// use wasmtime::component::{ResourceTable, Linker};
+/// use wasmtime_wasi_io::{IoView, add_to_linker_async};
+///
+/// struct MyState {
+///     table: ResourceTable,
+/// }
+///
+/// impl IoView for MyState {
+///     fn table(&mut self) -> &mut ResourceTable { &mut self.table }
+/// }
+/// let mut config = Config::new();
+/// config.async_support(true);
+/// let engine = Engine::new(&config).unwrap();
+/// let mut linker: Linker<MyState> = Linker::new(&engine);
+/// add_to_linker_async(&mut linker).unwrap();
+/// ```
+/// [`Store`]: wasmtime::Store
+/// [`Linker`]: wasmtime::component::Linker
+/// [`ResourceTable`]: wasmtime::component::ResourceTable
+///
+pub trait IoView: Send {
+    /// Yields mutable access to the internal resource management that this
+    /// context contains.
+    ///
+    /// Embedders can add custom resources to this table as well to give
+    /// resources to wasm as well.
+    fn table(&mut self) -> &mut ResourceTable;
+}
+impl<T: ?Sized + IoView> IoView for &mut T {
+    fn table(&mut self) -> &mut ResourceTable {
+        T::table(self)
+    }
+}
+impl<T: ?Sized + IoView> IoView for Box<T> {
+    fn table(&mut self) -> &mut ResourceTable {
+        T::table(self)
+    }
+}
+
+/// A small newtype wrapper which serves as the basis for implementations of
+/// `Host` WASI traits in this crate.
+///
+/// This type is used as the basis for the implementation of all `Host` traits
+/// generated by `bindgen!` for WASI interfaces.
+///
+/// You don't need to use this type if you are using the root
+/// [`add_to_linker_async`](crate::add_to_linker_async) in this crate.
+///
+/// If you're calling the `add_to_linker` functions generated by `bindgen!`
+/// from the [`bindings` module](crate::bindings), you'll want to create a
+/// value of this type in the closures added to a `Linker`.
+#[repr(transparent)]
+pub struct IoImpl<T>(pub T);
+impl<T: IoView> IoView for IoImpl<T> {
+    fn table(&mut self) -> &mut ResourceTable {
+        T::table(&mut self.0)
+    }
+}
 
 /// Add the wasi-io host implementation from this crate into the `linker`
 /// provided.
@@ -50,7 +122,7 @@ pub use async_trait::async_trait;
 /// ```
 /// use wasmtime::{Engine, Result, Store, Config};
 /// use wasmtime::component::{ResourceTable, Linker};
-/// use wasmtime_wasi_io::{IoView;
+/// use wasmtime_wasi_io::IoView;
 ///
 /// fn main() -> Result<()> {
 ///     let mut config = Config::new();
