--- crates/wasi-io/src/view.rs
@@ -1,71 +1,0 @@
-use wasmtime::component::ResourceTable;
-
-/// A trait which provides access to the [`ResourceTable`] inside the
-/// embedder's `T` of [`Store<T>`][`Store`].
-///
-/// This crate's WASI Host implementations depend on the contents of
-/// [`ResourceTable`]. The `T` type [`Store<T>`][`Store`] is defined in each
-/// embedding of Wasmtime. These implementations are connected to the
-/// [`Linker<T>`][`Linker`] by the
-/// [`add_to_linker_sync`](crate::add_to_linker_sync) and
-/// [`add_to_linker_async`](crate::add_to_linker_async) functions.
-///
-/// # Example
-///
-/// ```
-/// use wasmtime::component::{ResourceTable, Linker};
-/// use wasmtime_wasi_io::{IoView, add_to_linker};
-///
-/// struct MyState {
-///     table: ResourceTable,
-/// }
-///
-/// impl IoView for MyState {
-///     fn table(&mut self) -> &mut ResourceTable { &mut self.table }
-/// }
-///
-/// let mut linker: Linker<MyState> = Linker::new();
-/// add_to_linker_async(&mut linker);
-/// ```
-/// [`Store`]: wasmtime::Store
-/// [`Linker`]: wasmtime::component::Linker
-/// [`ResourceTable`]: wasmtime::component::ResourceTable
-///
-pub trait IoView: Send {
-    /// Yields mutable access to the internal resource management that this
-    /// context contains.
-    ///
-    /// Embedders can add custom resources to this table as well to give
-    /// resources to wasm as well.
-    fn table(&mut self) -> &mut ResourceTable;
-}
-impl<T: ?Sized + IoView> IoView for &mut T {
-    fn table(&mut self) -> &mut ResourceTable {
-        T::table(self)
-    }
-}
-impl<T: ?Sized + IoView> IoView for Box<T> {
-    fn table(&mut self) -> &mut ResourceTable {
-        T::table(self)
-    }
-}
-
-/// A small newtype wrapper which serves as the basis for implementations of
-/// `Host` WASI traits in this crate.
-///
-/// This type is used as the basis for the implementation of all `Host` traits
-/// generated by `bindgen!` for WASI interfaces.
-///
-/// You don't need to use this type if you are using the root
-/// [`add_to_linker_async`](crate::add_to_linker_async) in this crate.
-///
-/// If you're calling the `add_to_linker` functions generated by `bindgen!`
-/// from the [`bindings` module](crate::bindings), you'll want to create a
-/// value of this type in the closures added to a `Linker`.
-#[repr(transparent)]
-pub struct IoImpl<T>(pub T);
-impl<T: IoView> IoView for IoImpl<T> {
-    fn table(&mut self) -> &mut ResourceTable {
-        T::table(&mut self.0)
-    }
-}
