+++ crates/wasi/src/lib.rs
@@ -354,7 +354,7 @@ pub fn add_to_linker_with_options_async<T: WasiView>(
     options: &crate::bindings::LinkOptions,
 ) -> anyhow::Result<()> {
     let l = linker;
-    wasmtime_wasi_io::add_to_linker(l)?;
+    wasmtime_wasi_io::add_to_linker_async(l)?;
 
     let closure = type_annotate::<T, _>(|t| WasiImpl(IoImpl(t)));
